PE-78: fix smoke test v2
