+++ container/features/src/main/resources/features.xml
@@ -889,19 +889,6 @@
         <bundle>wrap:mvn:org.sblim.slp/sblimSLPClient/1.17</bundle>
         <bundle>wrap:mvn:org.sblim.wbem/sblimCIMClient/1.17</bundle>
         <bundle>mvn:org.opennms/opennms-vmware/${project.version}</bundle>
-        <!--bundle>mvn:org.opennms.features.config.dao/org.opennms.features.config.dao.api/${project.version}</bundle>
-        <bundle>mvn:org.opennms.features.config.dao/org.opennms.features.config.dao.impl/${project.version}</bundle>
-        <bundle>mvn:org.opennms.features.config.service/org.opennms.features.config.service.api/${project.version}</bundle>
-        <bundle>mvn:org.opennms.features.config.service/org.opennms.features.config.service.impl/${project.version}</bundle>
-        <bundle>mvn:org.json/json/${jsonVersion}</bundle>
-        <bundle dependency="true">mvn:com.fasterxml.jackson.core/jackson-databind/${jackson2Version}</bundle>
-        <bundle dependency="true">mvn:com.fasterxml.jackson.core/jackson-annotations/${jackson2Version}</bundle>
-        <bundle dependency="true">mvn:com.fasterxml.jackson.core/jackson-core/${jackson2Version}</bundle>
-        <bundle dependency="true">mvn:org.yaml/snakeyaml/${snakeyamlVersion}</bundle>
-        <bundle dependency="true">mvn:com.fasterxml.jackson.dataformat/jackson-dataformat-yaml/${jackson2Version}</bundle>
-        <bundle dependency="true">mvn:com.fasterxml.jackson.datatype/jackson-datatype-jsr310/${jackson2Version}</bundle>
-        <bundle>mvn:org.apache.ws.xmlschema/xmlschema-core/${xmlschemaVersion}</bundle>
-        <bundle>mvn:org.apache.ws.xmlschema/xmlschema-walker/${xmlschemaVersion}</bundle-->
     </feature>
     <feature name="opennms-poller-monitors-core" version="${project.version}" description="OpenNMS :: Poller :: Monitors :: Core">
         <feature>camel-core</feature>
