+++ integrations/opennms-vmware/src/main/java/org/opennms/netmgt/collectd/VmwareCollector.java
@@ -57,8 +57,8 @@ import org.opennms.netmgt.config.vmware.VmwareServer;
 import org.opennms.netmgt.config.vmware.vijava.Attrib;
 import org.opennms.netmgt.config.vmware.vijava.VmwareCollection;
 import org.opennms.netmgt.config.vmware.vijava.VmwareGroup;
-import org.opennms.netmgt.dao.VmwareConfigDao;
-import org.opennms.netmgt.dao.VmwareDatacollectionConfigDao;
+import org.opennms.netmgt.dao.vmware.VmwareConfigDao;
+import org.opennms.netmgt.dao.vmware.VmwareDatacollectionConfigDao;
 import org.opennms.netmgt.dao.api.NodeDao;
 import org.opennms.netmgt.model.OnmsNode;
 import org.opennms.netmgt.provision.service.vmware.VmwareImporter;
