+++ integrations/opennms-vmware/src/main/java/org/opennms/netmgt/dao/vmware/jaxb/VmwareDatacollectionConfigDaoJaxb.java
@@ -26,32 +26,31 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.dao.jaxb;
+package org.opennms.netmgt.dao.vmware.jaxb;
 
 import org.opennms.core.xml.AbstractJaxbConfigDao;
-import org.opennms.netmgt.config.vmware.cim.VmwareCimCollection;
-import org.opennms.netmgt.config.vmware.cim.VmwareCimDatacollectionConfig;
-import org.opennms.netmgt.dao.VmwareCimDatacollectionConfigDao;
+import org.opennms.netmgt.config.vmware.vijava.VmwareCollection;
+import org.opennms.netmgt.config.vmware.vijava.VmwareDatacollectionConfig;
+import org.opennms.netmgt.dao.vmware.VmwareDatacollectionConfigDao;
 import org.opennms.netmgt.rrd.RrdRepository;
 
 import java.io.File;
 import java.util.List;
 
 /**
- * The Class VmwareCimDatacollectionConfigDaoJaxb
+ * The Class DefaultVmwareDatacollectionConfigDao
  * <p/>
- * This class is used for accessing the Vmware Cim Data Collection configuration file
+ * This class is used for accessing the Vmware Data Collection configuration file
  *
  * @author Christian Pape <Christian.Pape@informatik.hs-fulda.de>
  */
-public class VmwareCimDatacollectionConfigDaoJaxb extends AbstractJaxbConfigDao<VmwareCimDatacollectionConfig, VmwareCimDatacollectionConfig> implements VmwareCimDatacollectionConfigDao {
-
+public class VmwareDatacollectionConfigDaoJaxb extends AbstractJaxbConfigDao<VmwareDatacollectionConfig, VmwareDatacollectionConfig> implements VmwareDatacollectionConfigDao {
 
     /**
      * Default constructor
      */
-    public VmwareCimDatacollectionConfigDaoJaxb() {
-        super(VmwareCimDatacollectionConfig.class, "Vmware Cim Data Collection Configuration");
+    public VmwareDatacollectionConfigDaoJaxb() {
+        super(VmwareDatacollectionConfig.class, "Vmware Data Collection Configuration");
     }
 
     /**
@@ -60,7 +59,7 @@ public class VmwareCimDatacollectionConfigDaoJaxb extends AbstractJaxbConfigDao<
      * @return the current config object
      */
     @Override
-    public VmwareCimDatacollectionConfig getConfig() {
+    public VmwareDatacollectionConfig getConfig() {
         return getContainer().getObject();
     }
 
@@ -72,7 +71,7 @@ public class VmwareCimDatacollectionConfigDaoJaxb extends AbstractJaxbConfigDao<
      * @return a custom object
      */
     @Override
-    public VmwareCimDatacollectionConfig translateConfig(VmwareCimDatacollectionConfig jaxbConfig) {
+    public VmwareDatacollectionConfig translateConfig(VmwareDatacollectionConfig jaxbConfig) {
         return jaxbConfig;
     }
 
@@ -80,20 +79,20 @@ public class VmwareCimDatacollectionConfigDaoJaxb extends AbstractJaxbConfigDao<
      * This method returns a subset of the configuration data for a given collection name.
      *
      * @param collectionName the collection's name
-     * @return the Cim collection object
+     * @return the collection object
      */
     @Override
-    public VmwareCimCollection getVmwareCimCollection(String collectionName) {
-        VmwareCimCollection[] collections = getConfig().getVmwareCimCollection();
-        VmwareCimCollection collection = null;
-        for (VmwareCimCollection coll : collections) {
+    public VmwareCollection getVmwareCollection(String collectionName) {
+        VmwareCollection[] collections = getConfig().getVmwareCollection();
+        VmwareCollection collection = null;
+        for (VmwareCollection coll : collections) {
             if (coll.getName().equalsIgnoreCase(collectionName)) {
                 collection = coll;
                 break;
             }
         }
         if (collection == null) {
-            throw new IllegalArgumentException("getVmwareCimCollection: collection name: "
+            throw new IllegalArgumentException("getVmwareCollection: collection name: "
                     + collectionName + " specified in collectd configuration not found in Vmware collection configuration.");
         }
         return collection;
@@ -122,7 +121,7 @@ public class VmwareCimDatacollectionConfigDaoJaxb extends AbstractJaxbConfigDao<
      * @return the step value
      */
     private int getStep(String cName) {
-        VmwareCimCollection collection = getVmwareCimCollection(cName);
+        VmwareCollection collection = getVmwareCollection(cName);
         if (collection != null) {
             return collection.getRrd().getStep();
         } else {
@@ -137,7 +136,7 @@ public class VmwareCimDatacollectionConfigDaoJaxb extends AbstractJaxbConfigDao<
      * @return the RRAs list
      */
     private List<String> getRRAList(String cName) {
-        VmwareCimCollection collection = getVmwareCimCollection(cName);
+        VmwareCollection collection = getVmwareCollection(cName);
         if (collection != null) {
             return collection.getRrd().getRraCollection();
         } else {
