+++ integrations/opennms-vmware/src/main/resources/META-INF/opennms/component-dao.xml
@@ -24,15 +24,15 @@
         <constructor-arg value="file:${opennms.home}/etc/vmware-config.xml"/>
     </bean>
 
-    <bean id="vmwareCimDatacollectionConfigDao" class="org.opennms.netmgt.dao.jaxb.VmwareCimDatacollectionConfigDaoJaxb">
+    <bean id="vmwareCimDatacollectionConfigDao" class="org.opennms.netmgt.dao.vmware.jaxb.VmwareCimDatacollectionConfigDaoJaxb">
         <property name="configResource" ref="vmwareCimDatacollectionConfigResourceLocation"/>
     </bean>
 
-    <bean id="vmwareDatacollectionConfigDao" class="org.opennms.netmgt.dao.jaxb.VmwareDatacollectionConfigDaoJaxb">
+    <bean id="vmwareDatacollectionConfigDao" class="org.opennms.netmgt.dao.vmware.jaxb.VmwareDatacollectionConfigDaoJaxb">
         <property name="configResource" ref="vmwareDatacollectionConfigResourceLocation"/>
     </bean>
 
-    <bean id="vmwareConfigDao" class="org.opennms.netmgt.dao.jaxb.VmwareConfigDaoJaxb">
+    <bean id="vmwareConfigDao" class="org.opennms.netmgt.dao.vmware.jaxb.VmwareConfigDaoJaxb">
         <property name="configResource" ref="vmwareConfigResourceLocation"/>
     </bean>
