+++ integrations/opennms-vmware/src/test/java/org/opennms/netmgt/collectd/VmwareCollectorComplianceTest.java
@@ -39,8 +39,8 @@ import java.util.Optional;
 import org.opennms.netmgt.collection.test.api.CollectorComplianceTest;
 import org.opennms.netmgt.config.vmware.VmwareServer;
 import org.opennms.netmgt.config.vmware.vijava.VmwareCollection;
-import org.opennms.netmgt.dao.VmwareConfigDao;
-import org.opennms.netmgt.dao.VmwareDatacollectionConfigDao;
+import org.opennms.netmgt.dao.vmware.VmwareConfigDao;
+import org.opennms.netmgt.dao.vmware.VmwareDatacollectionConfigDao;
 import org.opennms.netmgt.dao.api.NodeDao;
 import org.opennms.netmgt.dao.mock.MockTransactionTemplate;
 import org.opennms.netmgt.model.OnmsMetaData;
